Factor out State enums into separate files
This allows me to require them in multiple places, including test files
