+++ SessionStateXethruEnum.js
@@ -1,0 +1,13 @@
+'use strict';
+
+const XETHRU_STATE = {
+    BREATHING: "0",
+    MOVEMENT: "1",
+    MOVEMENT_TRACKING: "2",
+    NO_MOVEMENT: "3",
+    INITIALIZING: "4",
+    ERROR: "5",
+    UNKNOWN: "6"
+};
+
+module.exports = Object.freeze(XETHRU_STATE);
